Fixed Audio Extensions
Bug found where files with extensions in the path name threw an
extension. Extensions are now always removed from file names. You must
now ALWAYS specify the file extension of the audio file to be played.
Added a delegate for AudioPlayerQueue. It should now report when the
queue is empty/complete and when a sound is finished playing.