Build TinCan API from scratch..
I added the TinCan API. Had to build it from scratch into a static
library. Please DO NOT remove it and DO NOT remove CorePlot again.
Finally, there are two static libraries. The first is “combined”. It is
used for the simulator AND the phone. However for release builds, you
should use the “non-combined” version!